docs: move IVTime and DVTime to time.py
