+++ aer/experimentalist/experiment_environment/experiment.py
@@ -6,7 +6,7 @@ import aer.experimentalist.experiment_environment.experiment_config as config
 import aer.experimentalist.experimentalist_config as cfg
 from aer.variable import IVTrial
 from aer.variable.label import dv_labels, iv_labels
-from aer.variable.tinkerforge import DVTime, IVTime
+from aer.variable.time import DVTime, IVTime
 
 
 class Experiment:
