+++ aer/variable/time.py
@@ -1,5 +1,7 @@
 import time
 
+from aer.variable import DV, IV
+
 
 class VTime:
 
@@ -11,3 +13,57 @@ class VTime:
     # Resets reference time.
     def reset(self):
         self._t0 = time.time()
+
+
+class IVTime(IV, VTime):
+
+    _name = "time_IV"
+    _UID = ""
+    _variable_label = "Time"
+    _units = "s"
+    _priority = 0
+    _value_range = (0, 3600)
+    _value = 0
+
+    # Initializes reference time.
+    # The reference time usually denotes the beginning of an experiment trial.
+    def __init__(self, *args, **kwargs):
+        super(IVTime, self).__init__(*args, **kwargs)
+
+    # Waits until specified time has passed relative to reference time
+    def manipulate(self):
+
+        t_wait = self.get_value() - (time.time() - self._t0)
+        if t_wait <= 0:
+            return
+        else:
+            time.sleep(t_wait)
+
+    def disconnect(self):
+        pass
+
+
+class DVTime(DV, VTime):
+
+    _name = "time_DV"
+    _UID = ""
+    _variable_label = "Time"
+    _units = "s"
+    _priority = 0
+    _value_range = (0, 604800)  # don't record more than a week
+    _value = 0
+
+    _is_covariate = True
+
+    # Initializes reference time.
+    # The reference time usually denotes the beginning of an experiment trial.
+    def __init__(self, *args, **kwargs):
+        print(self._variable_label)
+        super(DVTime, self).__init__(*args, **kwargs)
+        print(self._variable_label)
+
+    # Measure number of seconds relative to reference time
+    def measure(self):
+
+        value = time.time() - self._t0
+        self.set_value(value)
