+++ aer/variable/tinkerforge.py
@@ -1,4 +1,3 @@
-import time
 from abc import abstractmethod
 
 from tinkerforge.bricklet_industrial_analog_out_v2 import BrickletIndustrialAnalogOutV2
@@ -9,7 +8,6 @@ from tinkerforge.bricklet_industrial_dual_analog_in_v2 import (
 from tinkerforge.ip_connection import IPConnection
 
 from aer.variable import DV, IV, Variable
-from aer.variable.time import VTime
 
 
 class TinkerforgeVariable(Variable):
@@ -71,34 +69,6 @@ class DVTF(DV, TinkerforgeVariable):
     def __init__(self, *args, **kwargs):
         DV.__init__(self, *args, **kwargs)
         TinkerforgeVariable.__init__(self, *args, **kwargs)
-
-
-class IVTime(IVTF, VTime):
-
-    _name = "time_IV"
-    _UID = ""
-    _variable_label = "Time"
-    _units = "s"
-    _priority = 0
-    _value_range = (0, 3600)
-    _value = 0
-
-    # Initializes reference time.
-    # The reference time usually denotes the beginning of an experiment trial.
-    def __init__(self, *args, **kwargs):
-        super(IVTime, self).__init__(*args, **kwargs)
-
-    # Waits until specified time has passed relative to reference time
-    def manipulate(self):
-
-        t_wait = self.get_value() - (time.time() - self._t0)
-        if t_wait <= 0:
-            return
-        else:
-            time.sleep(t_wait)
-
-    def disconnect(self):
-        pass
 
 
 class IVCurrent(IVTF):
@@ -189,32 +159,6 @@ class IVVoltage(IVTF):
 
     def clean_up(self):
         self.stop()
-
-
-class DVTime(DVTF, VTime):
-
-    _name = "time_DV"
-    _UID = ""
-    _variable_label = "Time"
-    _units = "s"
-    _priority = 0
-    _value_range = (0, 604800)  # don't record more than a week
-    _value = 0
-
-    _is_covariate = True
-
-    # Initializes reference time.
-    # The reference time usually denotes the beginning of an experiment trial.
-    def __init__(self, *args, **kwargs):
-        print(self._variable_label)
-        super(DVTime, self).__init__(*args, **kwargs)
-        print(self._variable_label)
-
-    # Measure number of seconds relative to reference time
-    def measure(self):
-
-        value = time.time() - self._t0
-        self.set_value(value)
 
 
 class DVCurrent(DVTF):
